pyproject.toml
[project]
name = "my-agent-project"
version = "0.1.0"
description = "我的第一个多Agent项目"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.5.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]
